+++ branches/modular-smart/src/navigators/smart/tom/core/timer/messages/RTLeaderChange.java
@@ -30,7 +30,7 @@ public class RTLeaderChange implements Serializable {
 
     public SignedObject[] proof; // Proofs for the new leader
     public int newLeader; // Replica ID of the new leader
-    public int start; // ID of the consensus to be started
+    public long start; // ID of the consensus to be started
 
     /**
      * Creates a new instance of RTLeaderChangeMessage
@@ -38,7 +38,7 @@ public class RTLeaderChange implements Serializable {
      * @param nl Replica ID of the new leader
      * @param start ID of the consensus to be started
      */
-    public RTLeaderChange(SignedObject[] proof, int nl, int start) {
+    public RTLeaderChange(SignedObject[] proof, int nl, long start) {
         this.proof = proof;
         this.newLeader = nl;
         this.start = start;
